requirements.txt
click==8.0.3
tqdm==4.62.3
pillow==8.4.0
numpy==1.21.5
opencv-python==4.5.4.60
ffmpeg-python==0.2.0
pyinstaller==4.7